requirements.txt
torch
torchvision
opencv-python
numpy
pillow
scikit-image
matplotlib
gradio
tqdm
torchinfo
datasets